Removed assert when unregistering an a writer that isn't registered.
This happens if OpenNI is initialized and shut down from within a static variable. There is no control if this code is called before or after the destruction of the LogData singleton.
If it's called after the destruction, the call for GetInstance() creates a new one, without any writers registered.
In any case, the result is the same - the writer isn't registered.

Signed-off-by: eddie.cohen <eddie.cohen@primesense.com>