Fix a memory corruption on Android.
OniFrameInternal struct is padded to 96 bytes (last data member ends after 92 bytes). But, when used in the pool, it is being derived by TInPool. For some reason, on Android, the first member of TInPool is placed at offset 92 (without padding). This causes corruption where performing memset according to sizeof(OniFrameInternal), which is 96.

Code was fixed to do this in the C++ way - initialize every member explicitly.